Moved VLIE type from crypto to types folder
