+++ src/frame/max.data.ts
@@ -1,5 +1,5 @@
 import {Bignum} from '../types/bignum';
-import {VLIE} from '../crypto/vlie';
+import {VLIE} from '../types/vlie';
 import {BaseFrame, FrameType} from './base.frame';
