+++ src/frame/new.connection.id.ts
@@ -1,7 +1,7 @@
 import {ConnectionID} from '../packet/header/header.properties';
 import {BaseFrame, FrameType} from './base.frame';
 import { Bignum } from '../types/bignum';
-import { VLIE } from '../crypto/vlie';
+import { VLIE } from '../types/vlie';
